requirements.txt
streamlit==1.56.0
langchain==1.2.15
langchain-openai==1.1.14
langchain-community==0.4.1
openai==2.32.0
PyMuPDF==1.27.2.2
faiss-cpu==1.13.2
python-dotenv==1.2.2
tiktoken==0.12.0
langchain-text-splitters==1.1.2
langchain-core==1.2.31
langchain-classic==1.0.4
